backend/pyproject.toml
[tool.poetry]
name = "swallow"
version = "0.1.0"
description = "just swallow"
authors = ["Aorts <t.sappalunn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
uvicorn = "^0.17.6"
mongoengine = "^0.24.1"
Pillow = "^9.0.1"
redis = "^4.2.0"
pydantic = {extras = ["email"], version = "^1.9.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
rq = "^1.10.1"
cryptography = "^36.0.2"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
